# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PesaGate Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  businessName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  email: string;
  businessName: string;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface ApiKey {
  id: number;
  keyName: string;
  publicKey: string;
  isActive: boolean;
  lastUsedAt?: string | null;
  createdAt: string;
}

export type ApiKeyWithSecret = ApiKey & {
  secretKey: string;
};

export interface CreateApiKeyRequest {
  keyName: string;
}

export type SettlementAccountAccountType =
  (typeof SettlementAccountAccountType)[keyof typeof SettlementAccountAccountType];

export const SettlementAccountAccountType = {
  till: "till",
  paybill: "paybill",
} as const;

export interface SettlementAccount {
  id: number;
  accountType: SettlementAccountAccountType;
  accountNumber: string;
  accountName: string;
  isDefault: boolean;
  isActive: boolean;
  createdAt: string;
}

export type AddSettlementAccountRequestAccountType =
  (typeof AddSettlementAccountRequestAccountType)[keyof typeof AddSettlementAccountRequestAccountType];

export const AddSettlementAccountRequestAccountType = {
  till: "till",
  paybill: "paybill",
} as const;

export interface AddSettlementAccountRequest {
  accountType: AddSettlementAccountRequestAccountType;
  accountNumber: string;
  accountName: string;
  isDefault?: boolean;
}

export interface STKPushRequest {
  /** Customer phone number (format 2547XXXXXXXX) */
  phoneNumber: string;
  /** @minimum 1 */
  amount: number;
  accountReference: string;
  transactionDesc: string;
  settlementAccountId?: number | null;
}

export interface STKPushResponse {
  checkoutRequestId: string;
  merchantRequestId: string;
  responseCode: string;
  responseDescription: string;
  customerMessage: string;
  transactionId: number;
}

export interface PaymentStatusResponse {
  checkoutRequestId: string;
  status: string;
  mpesaReceiptNumber?: string | null;
  amount: string;
  phoneNumber: string;
  createdAt: string;
  updatedAt: string;
}

export interface Transaction {
  id: number;
  checkoutRequestId?: string | null;
  mpesaReceiptNumber?: string | null;
  phoneNumber: string;
  amount: string;
  status: string;
  accountReference?: string | null;
  transactionDesc?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TransactionListResponse {
  transactions: Transaction[];
  total: number;
  page: number;
  limit: number;
}

export interface DashboardStats {
  totalTransactions: number;
  successfulTransactions: number;
  failedTransactions: number;
  pendingTransactions: number;
  totalVolume: string;
  todayVolume: string;
  todayTransactions: number;
  successRate: number;
}

export type ListTransactionsParams = {
  page?: number;
  limit?: number;
  status?: string;
};