# lib/api-zod/src/generated/types/addSettlementAccountRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PesaGate Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { AddSettlementAccountRequestAccountType } from "./addSettlementAccountRequestAccountType";

export interface AddSettlementAccountRequest {
  accountType: AddSettlementAccountRequestAccountType;
  accountNumber: string;
  accountName: string;
  isDefault?: boolean;
}