# lib/api-zod/src/generated/types/apiKey.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PesaGate Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKey {
  id: number;
  keyName: string;
  publicKey: string;
  isActive: boolean;
  lastUsedAt?: Date | null;
  createdAt: Date;
}