# lib/api-zod/src/generated/types/apiKeyWithSecret.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PesaGate Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKey } from "./apiKey";

export type ApiKeyWithSecret = ApiKey & {
  secretKey: string;
};